+++ Examples/Map+Positioning/Sources/LocationSourceType.swift
@@ -6,10 +6,6 @@
 //  Copyright © 2023 Wemap SAS. All rights reserved.
 //
 
-import WemapCoreSDK
-import WemapMapSDK
-import WemapPositioningSDK
-
 enum LocationSourceType: Int, CaseIterable {
     
     case vps,
@@ -20,11 +16,11 @@ enum LocationSourceType: Int, CaseIterable {
          
     var name: String {
         switch self {
-        case .simulator: return "Simulator"
-        case .polestar: return "Polestar"
-        case .systemDefault: return "System Default"
-        case .polestarEmulator: return "Polestar Emulator"
-        case .vps: return "VPS"
+        case .simulator: "Simulator"
+        case .polestar: "Polestar"
+        case .systemDefault: "System Default"
+        case .polestarEmulator: "Polestar Emulator"
+        case .vps: "VPS"
         }
     }
 }