+++ Examples/Sources/UIConstants.swift
@@ -1,0 +1,22 @@
+//
+//  UIConstants.swift
+//  Examples
+//
+//  Created by Evgenii Khrushchev on 18/12/2023.
+//  Copyright © 2023 Wemap SAS. All rights reserved.
+//
+
+import Foundation
+
+enum UIConstants {
+    
+    enum Delay {
+        static let short: TimeInterval = 5
+        static let long: TimeInterval = 10
+    }
+    
+    enum Inset {
+        static let top: CGFloat = -200
+        static let mid: CGFloat = -150
+    }
+}